Moved out HTTP functions from main to api module
